refactor: centralize language-specific parsing logic and rename parsing, diffing, and block parsing modules.
