+++ src/block_parser.rs
@@ -1,36 +1,8 @@
-mod bash;
-mod c;
-mod c_sharp;
-mod cpp;
-mod css;
-mod go;
-mod html;
-mod java;
-mod javascript;
-mod kotlin;
-mod makefile;
-mod markdown;
-mod php;
-mod python;
-mod ruby;
-mod rust;
-mod sql;
-mod swift;
-mod tag_parser;
-mod toml;
-mod tsx;
-mod typescript;
-mod xml;
-mod yaml;
-
 use crate::blocks::Block;
+use crate::language_parsers::{Comment, CommentsParser};
+use crate::tag_parser::{BlockTag, BlockTagParser, WinnowBlockTagParser};
 use std::collections::HashMap;
-use std::ffi::OsString;
 use std::ops::Range;
-use std::rc::Rc;
-use std::string::ToString;
-use tag_parser::{BlockTag, BlockTagParser, WinnowBlockTagParser};
-use tree_sitter::{Language, Node, Parser, Query, QueryCursor, StreamingIterator};
 
 /// Parses [`Blocks`] from a source code.
 pub trait BlocksParser {
@@ -40,166 +12,12 @@ pub trait BlocksParser {
     fn parse(&self, contents: &str) -> anyhow::Result<Vec<Block>>;
 }
 
-/// Parses comment strings from a source code.
-trait CommentsParser {
-    /// Returns a `Vec` of `Comment`s.
-    fn parse(&self, source_code: &str) -> anyhow::Result<Vec<Comment>>;
-}
-
-/// Returns a map of all available language parsers by their file extensions.
-pub fn language_parsers() -> anyhow::Result<HashMap<OsString, Rc<Box<dyn BlocksParser>>>> {
-    let bash_parser = Rc::new(Box::new(bash::parser()?) as Box<dyn BlocksParser>);
-    let c_parser = Rc::new(Box::new(c::parser()?) as Box<dyn BlocksParser>);
-    let c_sharp_parser = Rc::new(Box::new(c_sharp::parser()?) as Box<dyn BlocksParser>);
-    let cpp_parser = Rc::new(Box::new(cpp::parser()?) as Box<dyn BlocksParser>);
-    let css_parser = Rc::new(Box::new(css::parser()?) as Box<dyn BlocksParser>);
-    let go_parser = Rc::new(Box::new(go::parser()?) as Box<dyn BlocksParser>);
-    let html_parser = Rc::new(Box::new(html::parser()?) as Box<dyn BlocksParser>);
-    let java_parser = Rc::new(Box::new(java::parser()?) as Box<dyn BlocksParser>);
-    let js_parser = Rc::new(Box::new(javascript::parser()?) as Box<dyn BlocksParser>);
-    let kotlin_parser = Rc::new(Box::new(kotlin::parser()?) as Box<dyn BlocksParser>);
-    let makefile_parser = Rc::new(Box::new(makefile::parser()?) as Box<dyn BlocksParser>);
-    let markdown_parser = Rc::new(Box::new(markdown::parser()?) as Box<dyn BlocksParser>);
-    let php_parser = Rc::new(Box::new(php::parser()?) as Box<dyn BlocksParser>);
-    let python_parser = Rc::new(Box::new(python::parser()?) as Box<dyn BlocksParser>);
-    let ruby_parser = Rc::new(Box::new(ruby::parser()?) as Box<dyn BlocksParser>);
-    let rust_parser = Rc::new(Box::new(rust::parser()?) as Box<dyn BlocksParser>);
-    let sql_parser = Rc::new(Box::new(sql::parser()?) as Box<dyn BlocksParser>);
-    let swift_parser = Rc::new(Box::new(swift::parser()?) as Box<dyn BlocksParser>);
-    let toml_parser = Rc::new(Box::new(toml::parser()?) as Box<dyn BlocksParser>);
-    let typescript_parser = Rc::new(Box::new(typescript::parser()?) as Box<dyn BlocksParser>);
-    let typescript_tsx_parser = Rc::new(Box::new(tsx::parser()?) as Box<dyn BlocksParser>);
-    let xml_parser = Rc::new(Box::new(xml::parser()?) as Box<dyn BlocksParser>);
-    let yaml_parser = Rc::new(Box::new(yaml::parser()?) as Box<dyn BlocksParser>);
-    Ok(HashMap::from([
-        // <block affects="README.md:supported-grammar, src/blocks.rs:supported-extensions" keep-sorted="asc">
-        ("Makefile".into(), Rc::clone(&makefile_parser)),
-        ("bash".into(), Rc::clone(&bash_parser)),
-        ("c".into(), c_parser),
-        ("cc".into(), Rc::clone(&cpp_parser)),
-        ("cpp".into(), Rc::clone(&cpp_parser)),
-        ("cs".into(), c_sharp_parser),
-        ("css".into(), css_parser),
-        ("d.ts".into(), Rc::clone(&typescript_parser)),
-        ("go".into(), Rc::clone(&go_parser)),
-        ("go.mod".into(), Rc::clone(&go_parser)),
-        ("go.sum".into(), Rc::clone(&go_parser)),
-        ("go.work".into(), go_parser),
-        ("h".into(), cpp_parser),
-        ("htm".into(), Rc::clone(&html_parser)),
-        ("html".into(), html_parser),
-        ("java".into(), java_parser),
-        ("js".into(), Rc::clone(&js_parser)),
-        ("jsx".into(), js_parser),
-        ("kt".into(), Rc::clone(&kotlin_parser)),
-        ("kts".into(), kotlin_parser),
-        ("makefile".into(), Rc::clone(&makefile_parser)),
-        ("markdown".into(), Rc::clone(&markdown_parser)),
-        ("md".into(), markdown_parser),
-        ("mk".into(), makefile_parser),
-        ("php".into(), Rc::clone(&php_parser)),
-        ("phtml".into(), php_parser),
-        ("py".into(), Rc::clone(&python_parser)),
-        ("pyi".into(), python_parser),
-        ("rb".into(), ruby_parser),
-        ("rs".into(), rust_parser),
-        ("sh".into(), bash_parser),
-        ("sql".into(), sql_parser),
-        ("swift".into(), swift_parser),
-        ("toml".into(), toml_parser),
-        ("ts".into(), typescript_parser),
-        ("tsx".into(), typescript_tsx_parser),
-        ("xml".into(), xml_parser),
-        ("yaml".into(), Rc::clone(&yaml_parser)),
-        ("yml".into(), yaml_parser),
-        // </block>
-    ]))
-}
-
-type CaptureProcessor = Box<dyn Fn(usize, &str, &Node) -> anyhow::Result<Option<String>>>;
-
-struct TreeSitterCommentsParser {
-    language: Language,
-    queries: Vec<(Query, Option<CaptureProcessor>)>,
-}
-
-impl TreeSitterCommentsParser {
-    fn new(language: Language, queries: Vec<(Query, Option<CaptureProcessor>)>) -> Self {
-        Self { language, queries }
-    }
-}
-
-impl CommentsParser for TreeSitterCommentsParser {
-    fn parse(&self, source_code: &str) -> anyhow::Result<Vec<Comment>> {
-        let mut parser = Parser::new();
-        parser
-            .set_language(&self.language)
-            .expect("Error setting Tree-sitter language");
-        let tree = parser.parse(source_code, None).unwrap();
-        let root_node = tree.root_node();
-
-        let mut blocks = vec![];
-        for (query, post_processor) in self.queries.iter() {
-            let mut query_cursor = QueryCursor::new();
-            let mut matches = query_cursor.matches(query, root_node, source_code.as_bytes());
-            while let Some(query_match) = matches.next() {
-                for capture in query_match.captures {
-                    let node = capture.node;
-                    let start_line = node.start_position().row + 1; // Convert to 1-based indexing
-                    let start_position = node.start_byte();
-                    let end_position = node.end_byte();
-                    let comment_text = &source_code[node.start_byte()..node.end_byte()];
-                    if let Some(processor) = post_processor {
-                        if let Some(out) = processor(capture.index as usize, comment_text, &node)? {
-                            blocks.push(Comment {
-                                source_line_number: start_line,
-                                source_start_position: start_position,
-                                source_end_position: end_position,
-                                comment_text: out,
-                            });
-                        }
-                    } else {
-                        blocks.push(Comment {
-                            source_line_number: start_line,
-                            source_start_position: start_position,
-                            source_end_position: end_position,
-                            comment_text: comment_text.to_string(),
-                        });
-                    }
-                }
-            }
-        }
-
-        blocks.sort_by(|comment1, comment2| {
-            comment1
-                .source_start_position
-                .cmp(&comment2.source_start_position)
-        });
-        Ok(blocks)
-    }
-}
-
-struct BlocksFromCommentsParser<C: CommentsParser> {
+pub struct BlocksFromCommentsParser<C: CommentsParser> {
     comments_parser: C,
 }
 
-#[derive(Debug, PartialEq)]
-struct Comment {
-    // 1-based line number of the start of this comment in the source.
-    source_line_number: usize,
-    // Byte offset (i.e. position) of the start of this comment in the source.
-    source_start_position: usize,
-    // Byte offset (i.e. position) of the end of this comment in the source.
-    source_end_position: usize,
-    // The `comment_string` is expected to be the content of the comment with all language specific
-    // comment symbols like "//", "/**", "#", etc replaced with the corresponding number of
-    // whitespaces ("  " for "//", "   " for "/**", etc.) so that the length of the comment is
-    // preserved.
-    comment_text: String,
-}
-
 impl<C: CommentsParser> BlocksFromCommentsParser<C> {
-    fn new(comments_parser: C) -> Self {
+    pub(crate) fn new(comments_parser: C) -> Self {
         Self { comments_parser }
     }
 
@@ -328,133 +146,16 @@ impl<'c> BlockBuilder<'c> {
     }
 }
 
-/// C-style comments parser for a query that returns both line and block comments.
-fn c_style_comments_parser(language: Language, query: Query) -> TreeSitterCommentsParser {
-    TreeSitterCommentsParser::new(
-        language,
-        vec![(
-            query,
-            Some(Box::new(|_, comment, _node| {
-                let result = if comment.starts_with("//") {
-                    comment.replacen("//", "  ", 1)
-                } else {
-                    c_style_multiline_comment_processor(comment)
-                };
-                Ok(Some(result))
-            })),
-        )],
-    )
-}
-
-/// C-style comments parser for the separate line and block comment queries.
-fn c_style_line_and_block_comments_parser(
-    language: Language,
-    line_comment_query: Query,
-    block_comment_query: Query,
-) -> TreeSitterCommentsParser {
-    TreeSitterCommentsParser::new(
-        language,
-        vec![
-            (
-                line_comment_query,
-                Some(Box::new(|_, comment, _node| {
-                    Ok(Some(comment.replacen("//", "  ", 1)))
-                })),
-            ),
-            (
-                block_comment_query,
-                Some(Box::new(|_, comment, _node| {
-                    Ok(Some(c_style_multiline_comment_processor(comment)))
-                })),
-            ),
-        ],
-    )
-}
-
-/// Python-style comments parser.
-fn python_style_comments_parser(
-    language: Language,
-    comment_query: Query,
-) -> TreeSitterCommentsParser {
-    TreeSitterCommentsParser::new(
-        language,
-        vec![(
-            comment_query,
-            Some(Box::new(|_, comment, _node| {
-                Ok(Some(comment.replacen("#", " ", 1)))
-            })),
-        )],
-    )
-}
-
-/// XML-style comments parser.
-fn xml_style_comments_parser(language: Language, comment_query: Query) -> TreeSitterCommentsParser {
-    TreeSitterCommentsParser::new(
-        language,
-        vec![(
-            comment_query,
-            Some(Box::new(|_, comment, _node| {
-                let open_idx = comment.find("<!--").expect("open comment tag is expected");
-                let close_idx = comment.rfind("-->").expect("close comment tag is expected");
-                let mut result = String::with_capacity(comment.len());
-                result.push_str(&comment[..open_idx]);
-                // Replace "<!--" with spaces.
-                result.push_str("    ");
-                result.push_str(&comment[open_idx + 4..close_idx]);
-                // Replace "-->" with spaces.
-                result.push_str("   ");
-                result.push_str(&comment[close_idx + 3..]);
-                Ok(Some(result))
-            })),
-        )],
-    )
-}
-
-fn c_style_multiline_comment_processor(comment: &str) -> String {
-    let mut result = String::with_capacity(comment.len());
-    let open_idx = comment.find("/*").expect("expected '/*' in a comment");
-    let close_idx = comment.rfind("*/").expect("expected '*/' in a comment");
-    // Add everything before the "/*"
-    result.push_str(&comment[..open_idx]);
-    // Replace "/*" with spaces.
-    result.push_str("  ");
-    let content = &comment[open_idx + 2..close_idx];
-    for line in content.split_inclusive('\n') {
-        let mut decorative_star_found = false;
-
-        // Find the index of the first non-whitespace character
-        if let Some(first_non_whitespace_idx) = line.find(|c: char| !c.is_whitespace()) {
-            // Check if that first non-whitespace character is a '*'
-            if line[first_non_whitespace_idx..].starts_with('*') {
-                decorative_star_found = true;
-                // Add leading whitespace.
-                result.push_str(&line[..first_non_whitespace_idx]);
-                // Replace "*" with a space.
-                result.push(' ');
-                // Add the rest of the line.
-                result.push_str(&line[first_non_whitespace_idx + 1..]);
-            }
-        }
-        if !decorative_star_found {
-            // Not a decorative '*', or all whitespace. Add unchanged.
-            result.push_str(line);
-        }
-    }
-    // Replace "*/" with spaces.
-    result.push_str("  ");
-    // Add everything after the "*/".
-    result.push_str(&comment[close_idx + 2..]);
-
-    result
-}
-
 #[cfg(test)]
 mod tests {
-    use super::*;
-    use crate::test_utils;
+    use crate::block_parser::BlocksParser;
+    use crate::blocks::Block;
+    use crate::{language_parsers, test_utils};
+    use std::collections::HashMap;
 
     fn create_parser() -> impl BlocksParser {
-        rust::parser().unwrap()
+        // Reuse existing real blocks parser.
+        language_parsers::rust::parser().unwrap()
     }
 
     #[test]
