+++ src/language_parsers/kotlin.rs
@@ -1,6 +1,5 @@
-use crate::parsers::{
-    BlocksFromCommentsParser, BlocksParser, CommentsParser, c_style_line_and_block_comments_parser,
-};
+use crate::block_parser::{BlocksFromCommentsParser, BlocksParser};
+use crate::language_parsers::{CommentsParser, c_style_line_and_block_comments_parser};
 use tree_sitter::Query;
 
 /// Returns a [`BlocksParser`] for Kotlin.
@@ -23,7 +22,7 @@ fn comments_parser() -> anyhow::Result<impl CommentsParser> {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::parsers::Comment;
+    use crate::language_parsers::Comment;
 
     #[test]
     fn parses_kotlin_comments_correctly() -> anyhow::Result<()> {
