+++ src/language_parsers/rust.rs
@@ -1,11 +1,11 @@
-use crate::parsers::{
-    BlocksFromCommentsParser, BlocksParser, CommentsParser, TreeSitterCommentsParser,
-    c_style_multiline_comment_processor,
+use crate::block_parser::{BlocksFromCommentsParser, BlocksParser};
+use crate::language_parsers::{
+    CommentsParser, TreeSitterCommentsParser, c_style_multiline_comment_processor,
 };
 use tree_sitter::Query;
 
 /// Returns a [`BlocksParser`] for Rust.
-pub(super) fn parser() -> anyhow::Result<impl BlocksParser> {
+pub(crate) fn parser() -> anyhow::Result<impl BlocksParser> {
     Ok(BlocksFromCommentsParser::new(comments_parser()?))
 }
 
@@ -44,7 +44,7 @@ fn comments_parser() -> anyhow::Result<impl CommentsParser> {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::parsers::Comment;
+    use crate::language_parsers::Comment;
 
     #[test]
     fn parses_comments_correctly() -> anyhow::Result<()> {
